pyproject.toml
[project]
name = "svgllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.0",
    "beautifulsoup4>=4.13.4",
    "cairosvg>=2.8.2",
    "numpy>=2.3.2",
    "peft>=0.17.0",
    "pillow>=11.3.0",
    "py7zr==0.20.8",
    "requests>=2.32.4",
    "scikit-image>=0.25.2",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.0",
]
